* xim/canddisp.cpp   - (candwin_read_cb): Suppress warnings by casting. * xim/compose.cpp   - (XimIM::get_mb_string, XimIM::parse_compose_line, KeySymToUcs4,      mb_string_to_utf8): Ditto. * xim/connection.cpp   - (XConnection::readToBuf): Ditto. * xim/convdisp.cpp   - (PeWin::draw_char, PeWin::set_back, PeLineWin::get_char_width,      ConvdispOv::check_atr, ConvdispOv::layoutCharEnt,      ConvdispOs::compose_preedit_array,      ConvdispOs::compose_feedback_array): Ditto. * xim/helper.cpp   - (send_im_list, helper_str_parse): Ditto. * xim/locale.cpp   - (valid_locales, all_locales): Ditto. * xim/main.cpp   - (add_window_watch, remove_window_watch, findWindowIf, notifyLocale,      init_supported_locales): Ditto. * xim/ximic.cpp   - (icxatr::icxatr, XimIC::send_key_event, XimIC::get_ic_atr,      XimIC::reset_ic, XimIC::onSendPacket): Ditto. * xim/ximpacket.cpp   - (TxString::TxString): Ditto. * xim/ximserver.cpp   - (XimServer::strToUstring,      InputContext::get_caret_state_label_from_prop_list,      InputContext::update_prop_list, keyState::check_key): Ditto. * xim/ximtrans.cpp   - (XIMATTRIBUTE::write_imattr_to_packet,      XICATTRIBUTE::write_icattr_to_packet,      Connection::push_error_packet): Ditto.

+++ xim/canddisp.cpp
@@ -293,7 +293,7 @@ static void candwin_read_cb(int fd, int /* ev */)
     char buf[1024];
     int n;
 
-    n = read(fd, buf, 1024 - 1);
+    n = static_cast<int>(read(fd, buf, 1024 - 1));
     if (n == 0) {
 	terminate_canddisp_connection();
 	return;
